+++ integreat_cms/api/v3/regions.py
@@ -1,6 +1,7 @@
 """
 This module includes functions related to the regions API endpoint.
 """
+from django.conf import settings
 from django.http import JsonResponse
 
 from ...cms.models import Region
@@ -32,6 +33,8 @@ def transform_region(region):
         "push_notifications": region.push_notifications_enabled,
         "longitude": region.longitude,
         "latitude": region.latitude,
+        # TODO: Store bounding box in region model
+        "bounding_box": settings.DEFAULT_BOUNDING_BOX,
         "aliases": region.aliases,
         "tunews": region.tunews_enabled,
     }